Tests, fixtures, split tests between storefront/dashboard, redirect url with inproper slug
